Create TrailingStopLoss.py
Create an algorithm on QuantConnect that does the following:

Backtesting Period: Jan 1st, 2018 to Jan 1st 2019

Starting Cash: $100,000
Implemented a trailing stop loss, which is a StopMarketOrder with a Stop price that is continually updated and moved higher if the underlying security price moves higher. This allows you to have an adjusted risk profile, where you gradually move your stop price in concert with the share price, instead of a static stop price set at the beginning of the order and never changed.